+++ src/User/RegisterUserCommand.php
@@ -61,6 +61,9 @@ final class RegisterUserCommand
     /** @var string|null */
     private $legalIdentifier;
 
+    /** @var string|null */
+    private $loyaltyIdentifier;
+
     /** @return string|null */
     public function getPhone(): ?string
     {
@@ -359,6 +362,24 @@ final class RegisterUserCommand
     public function setLegalIdentifier(?string $legalIdentifier): self
     {
         $this->legalIdentifier = $legalIdentifier;
+
+        return $this;
+    }
+
+    /** @return string|null */
+    public function getLoyaltyIdentifier(): ?string
+    {
+        return $this->loyaltyIdentifier;
+    }
+
+    /**
+     * @param string|null $loyaltyIdentifier
+     *
+     * @return $this
+     */
+    public function setLoyaltyIdentifier(?string $loyaltyIdentifier): self
+    {
+        $this->loyaltyIdentifier = $loyaltyIdentifier;
 
         return $this;
     }